hal/armv7a: apply new structure of hal interface
JIRA: PD-210
