+++ hal/armv7a/imx6ull/Makefile
@@ -6,4 +6,4 @@
 
 OBJS += $(addprefix $(PREFIX_O)hal/armv7a/imx6ull/, _init.o console.o imx6ull.o timer.o interrupts.o)
 
-$(PREFIX_O)hal/armv7a/imx6ull/_init.o: hal/armv7a/_hal.S hal/armv7a/_interrupts.S
+$(PREFIX_O)hal/armv7a/imx6ull/_init.o: hal/armv7a/_armv7a.S hal/armv7a/_interrupts.S
