+++ hal/armv7a/imx6ull/console.c
@@ -14,8 +14,7 @@
  */
 
 #include "../../console.h"
-#include "cpu.h"
-#include "syspage.h"
+#include "../../cpu.h"
 
 #define UART uart1
