+++ hal/armv7a/imx6ull/interrupts.c
@@ -13,16 +13,49 @@
  * %LICENSE%
  */
 
-#include "cpu.h"
 #include "pmap.h"
-#include "spinlock.h"
-#include "interrupts.h"
+#include "../../cpu.h"
+#include "../../interrupts.h"
 
 #include "../../proc/userintr.h"
 #include "../../include/errno.h"
 
 #define SIZE_INTERRUPTS 159
 #define SIZE_HANDLERS   4
+
+
+#define _intr_add(list, t) \
+	do { \
+		if (t == NULL) \
+			break; \
+		if (*list == NULL) { \
+			t->next = t; \
+			t->prev = t; \
+			(*list) = t; \
+			break; \
+		} \
+		t->prev = (*list)->prev; \
+		(*list)->prev->next = t; \
+		t->next = (*list); \
+		(*list)->prev = t; \
+	} while (0)
+
+
+#define _intr_remove(list, t) \
+	do { \
+		if (t == NULL) \
+			break; \
+		if ((t->next == t) && (t->prev == t)) \
+			(*list) = NULL; \
+		else { \
+			t->prev->next = t->next; \
+			t->next->prev = t->prev; \
+			if (t == (*list)) \
+				(*list) = t->next; \
+		} \
+		t->next = NULL; \
+		t->prev = NULL; \
+	} while (0)
 
 
 enum { /* 1024 reserved */ ctlr = 0x400, typer, iidr, /* 29 reserved */ igroupr0 = 0x420, /* 16 registers */
