+++ hal/armv7a/zynq7000/interrupts.c
@@ -13,12 +13,44 @@
  * %LICENSE%
  */
 
-#include "cpu.h"
-#include "pmap.h"
-#include "spinlock.h"
-#include "interrupts.h"
+#include "../../cpu.h"
+#include "../../spinlock.h"
+#include "../../interrupts.h"
 
 #include "../../proc/userintr.h"
+
+#define _intr_add(list, t) \
+	do { \
+		if (t == NULL) \
+			break; \
+		if (*list == NULL) { \
+			t->next = t; \
+			t->prev = t; \
+			(*list) = t; \
+			break; \
+		} \
+		t->prev = (*list)->prev; \
+		(*list)->prev->next = t; \
+		t->next = (*list); \
+		(*list)->prev = t; \
+	} while (0)
+
+
+#define _intr_remove(list, t) \
+	do { \
+		if (t == NULL) \
+			break; \
+		if ((t->next == t) && (t->prev == t)) \
+			(*list) = NULL; \
+		else { \
+			t->prev->next = t->next; \
+			t->next->prev = t->prev; \
+			if (t == (*list)) \
+				(*list) = t->next; \
+		} \
+		t->next = NULL; \
+		t->prev = NULL; \
+	} while (0)
 
 #define SIZE_INTERRUPTS 95
 #define SIZE_HANDLERS   4
